requirements.txt
torch==2.1.0
transformers==4.44.0
accelerate==0.32.0
peft==0.12.0
numpy==1.26.0
pandas==2.2.0
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.0
plotly==5.22.0
wandb==0.17.0
nixtla==0.5.0
neuralforecast==1.7.0
pyyaml==6.0.1
tqdm==4.66.0
